requirements.txt
annotated-types==0.7.0
anyio==4.10.0
bcrypt==4.3.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
Deprecated==1.2.18
ecdsa==0.19.1
fastapi==0.116.2
google-auth==2.40.3
google-auth-oauthlib==1.2.2
greenlet==3.2.4
h11==0.16.0
idna==3.10
limits==5.5.0
oauthlib==3.3.1
packaging==25.0
passlib==1.7.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.9
pydantic_core==2.33.2
python-jose==3.5.0
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.36.0
wrapt==1.17.3
